Fixed bounds-checking problem
the problem with checking the high byte of mscr and mcolor by
themselves, is that they don't include the Y-index.  so I had
to add some math before each check to compensate for that.
This fixes the problem with the keyboard becoming unusable --
when a raindrop fell off the screen, the color part would
fall into the address space of the CIA chips or something.